+++ projects/quotes-widget/src/app/sm-quotes/quote-list/quote-list.component.ts
@@ -25,9 +25,25 @@ export class QuoteListComponent implements OnInit {
     this.subscription
       .add(
         quoteService.getQuotes()
-          .subscribe(data => {
-            this.listOfQuotes = data;
-          }))
+          .subscribe(
+            data => this.listOfQuotes = data,
+            error => this.listOfQuotes = [
+              {
+                id: 999,
+                speakers: [
+                  {
+                    person: "Steven Meehan",
+                    words: "In life, journeys are seldom easy. Despite the internet invading every aspect of our lives, data sometimes gets lost. While I’m disappointed that my curated quotes are not being rendered, at least the articles are still available. My development team is pulling every log file available in their search for the problem. With fingers crossed, I’m hoping the required fix is simple so my team restores full functionality soon.",
+                    order: 0
+                  }
+                ],
+                source: {
+                  story: "stevenmeehan.com",
+                  series: ""
+                }
+              }
+            ] 
+          ))
       .add(
         appConfigService.GetNumberOfCharactersToDisplay()
           .subscribe(result => {