+++ changelog.md
@@ -2,7 +2,12 @@
 
 ## vNext
 
-## 1.0.0 (1)
+## 1.1.0
+
+- Added error handeling for web requests Quptes and Progress Projects
+- Updated the Base URL for Works in Progress 
+
+## 1.0.0
 
 - Converted the Quotes and Progress widget to Angular 11, and consolidated
 them into a single workspace